Start implementing Isotropic.
Also move Single to its own private module file.
